+++ runtimes-test-client/test.cjs
@@ -1,0 +1,3 @@
+const conditions = require('@sanity/runtimes-test-client/supports-conditions')
+
+console.log(conditions)