+++ runtimes-test-client/test.js
@@ -1,0 +1,3 @@
+import conditions from '@sanity/runtimes-test-client/supports-conditions'
+
+console.log(conditions)