+++ runtimes/bun-script/test.expected.js
@@ -1,6 +1,6 @@
 import { write } from 'bun'
-import createClient, { entry } from 'skunkworks-test-client'
-import conditions from 'skunkworks-test-client/supports-conditions'
+import createClient, { entry } from '@sanity/runtimes-test-client'
+import conditions from '@sanity/runtimes-test-client/supports-conditions'
 
 import { projectId, dataset, apiVersion, query } from './config.js'