+++ runtimes/deno-cdn-script/test.ts
@@ -1,7 +1,7 @@
 import createClientExpected, {
   entry,
-} from 'https://esm.sh/skunkworks-test-client'
-import conditions from 'https://esm.sh/skunkworks-test-client/supports-conditions'
+} from 'https://esm.sh/@sanity/runtimes-test-client'
+import conditions from 'https://esm.sh/@sanity/runtimes-test-client/supports-conditions'
 import { createClient as createClientActual } from 'https://esm.sh/@sanity/client'
 
 const projectId = '81pocpw8'