+++ skunkworks-test-client/test.cjs
@@ -1,3 +1,0 @@
-const conditions = require('skunkworks-test-client/supports-conditions')
-
-console.log(conditions)