+++ skunkworks-test-client/test.js
@@ -1,3 +1,0 @@
-import conditions from 'skunkworks-test-client/supports-conditions'
-
-console.log(conditions)